gestures: Explicitly instantiate ZOMBULL in enums.js
Explicitly instantiate the ZOMBULL "namespace" in enums.js, as scripts must
be loaded in the exact order in which they are specified, i.e. enums.js
must be loaded first, otherwise the enums won't be accessible to others.

Opportunsitically make ZOMBULL fully global in anticipation of moving the
background thread to a service worker (which doesnt have an associated
window object).