gestures: Expose ActionTargets directly to content scripts and extension pages
Consume ZOMBULL.ActionTargets directly in gestures.js instead of having the
dictionary passed back from the background script in anticipation of
removing the use of long-lived port connections (which is more or less
necessary to move the background page to a service worker).

Note, defaultsoptions.js needs to be imported in both the content script
and the newtab page (the extension's scripts aren't injected into the
newtab page, i.e. they must be explicitly loaded).