gestures: Increase indentation of background script
Increase the indentation of the background script to isolate upcoming
changes for the conversion to manifest v3.  Moving the background page to
a service worker will require using importScript(), which in turn will
require checking for the existence of importScript() (because apparently
the service worker script is loaded in multiple contexts?).